requirements.txt
python-telegram-bot==20.8
python-dotenv
asyncio
nest_asyncio
